Split EntityUpdater into separate code files
